refactor politeness strats tests to include DB mode
